requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
jsonfield==3.1.0
phonenumbers==8.12.22
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
